+++ services/sharedurls/service.go
@@ -1,15 +1,11 @@
-package protocol
+package sharedurls
 
 import (
-	"bytes"
-	"encoding/base64"
 	"fmt"
 	"regexp"
-	"strings"
 
 	"github.com/golang/protobuf/proto"
-
-	"github.com/andybalholm/brotli"
+	"github.com/pkg/errors"
 
 	"github.com/status-im/status-go/api/multiformat"
 	gocommon "github.com/status-im/status-go/common"
@@ -22,36 +18,6 @@ import (
 	"github.com/status-im/status-go/protocol/requests"
 	"github.com/status-im/status-go/services/utils"
 )
-
-type CommunityURLData struct {
-	DisplayName  string   `json:"displayName"`
-	Description  string   `json:"description"`
-	MembersCount uint32   `json:"membersCount"`
-	Color        string   `json:"color"`
-	TagIndices   []uint32 `json:"tagIndices"`
-	CommunityID  string   `json:"communityId"`
-}
-
-type CommunityChannelURLData struct {
-	Emoji       string `json:"emoji"`
-	DisplayName string `json:"displayName"`
-	Description string `json:"description"`
-	Color       string `json:"color"`
-	ChannelUUID string `json:"channelUuid"`
-}
-
-type ContactURLData struct {
-	DisplayName string `json:"displayName"`
-	Description string `json:"description"`
-	PublicKey   string `json:"publicKey"`
-}
-
-type URLDataResponse struct {
-	Community *CommunityURLData        `json:"community"`
-	Channel   *CommunityChannelURLData `json:"channel"`
-	Contact   *ContactURLData          `json:"contact"`
-	Shard     *messagingtypes.Shard    `json:"shard,omitempty"`
-}
 
 const baseShareURL = "https://status.app"
 const userPath = "u#"
@@ -69,6 +35,28 @@ const sharedURLChannelPrefixWithData = baseShareURL + "/" + channelPath
 const channelUUIDRegExp = "^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$"
 
 var channelRegExp = regexp.MustCompile(channelUUIDRegExp)
+
+var (
+	ErrContactNotFound = errors.New("contact not found")
+)
+
+type Service struct {
+	provider DataProvider
+}
+
+func NewService(provider DataProvider) *Service {
+	return &Service{
+		provider,
+	}
+}
+
+func (s *Service) Start() error {
+	return nil
+}
+
+func (s *Service) Stop() error {
+	return nil
+}
 
 func decodeCommunityID(serialisedPublicKey string) (string, error) {
 	deserializedCommunityID, err := multiformat.DeserializeCompressedKey(serialisedPublicKey)
@@ -92,7 +80,7 @@ func deserializePublicKey(compressedKey string) (types.HexBytes, error) {
 	return utils.DeserializePublicKey(compressedKey)
 }
 
-func (m *Messenger) ShareCommunityURLWithChatKey(communityID types.HexBytes) (string, error) {
+func (m *Service) ShareCommunityURLWithChatKey(communityID types.HexBytes) (string, error) {
 	shortKey, err := serializePublicKey(communityID)
 	if err != nil {
 		return "", err
@@ -115,7 +103,7 @@ func parseCommunityURLWithChatKey(urlData string) (*URLDataResponse, error) {
 	}, nil
 }
 
-func (m *Messenger) prepareEncodedCommunityData(community *communities.Community) (string, string, error) {
+func (m *Service) prepareEncodedCommunityData(community *communities.Community) (string, string, error) {
 	communityProto := &protobuf.Community{
 		DisplayName:  community.Identity().DisplayName,
 		Description:  community.DescriptionText(),
@@ -152,8 +140,8 @@ func (m *Messenger) prepareEncodedCommunityData(community *communities.Community
 	return encodedData, shortKey, nil
 }
 
-func (m *Messenger) ShareCommunityURLWithData(communityID types.HexBytes) (string, error) {
-	community, err := m.GetCommunityByID(communityID)
+func (m *Service) ShareCommunityURLWithData(communityID types.HexBytes) (string, error) {
+	community, err := m.provider.GetCommunityByID(communityID)
 	if err != nil {
 		return "", err
 	}
@@ -209,7 +197,7 @@ func parseCommunityURLWithData(data string, chatKey string) (*URLDataResponse, e
 	}, nil
 }
 
-func (m *Messenger) ShareCommunityChannelURLWithChatKey(request *requests.CommunityChannelShareURL) (string, error) {
+func (m *Service) ShareCommunityChannelURLWithChatKey(request *requests.CommunityChannelShareURL) (string, error) {
 	if err := request.Validate(); err != nil {
 		return "", err
 	}
@@ -258,7 +246,7 @@ func parseCommunityChannelURLWithChatKey(channelID string, publicKey string) (*U
 	}, nil
 }
 
-func (m *Messenger) prepareEncodedCommunityChannelData(community *communities.Community, channel *protobuf.CommunityChat, channelID string) (string, string, error) {
+func (m *Service) prepareEncodedCommunityChannelData(community *communities.Community, channel *protobuf.CommunityChat, channelID string) (string, string, error) {
 	communityProto := &protobuf.Community{
 		DisplayName:  community.Identity().DisplayName,
 		Description:  community.DescriptionText(),
@@ -303,7 +291,7 @@ func (m *Messenger) prepareEncodedCommunityChannelData(community *communities.Co
 	return encodedData, shortKey, nil
 }
 
-func (m *Messenger) ShareCommunityChannelURLWithData(request *requests.CommunityChannelShareURL) (string, error) {
+func (m *Service) ShareCommunityChannelURLWithData(request *requests.CommunityChannelShareURL) (string, error) {
 	if err := request.Validate(); err != nil {
 		return "", err
 	}
@@ -317,7 +305,7 @@ func (m *Messenger) ShareCommunityChannelURLWithData(request *requests.Community
 		return "", fmt.Errorf("channelID should be UUID, got %s", gocommon.TruncateWithDot(request.ChannelID))
 	}
 
-	community, err := m.GetCommunityByID(request.CommunityID)
+	community, err := m.provider.GetCommunityByID(request.CommunityID)
 	if err != nil {
 		return "", err
 	}
@@ -385,7 +373,7 @@ func parseCommunityChannelURLWithData(data string, chatKey string) (*URLDataResp
 	}, nil
 }
 
-func (m *Messenger) ShareUserURLWithChatKey(contactID string) (string, error) {
+func (m *Service) ShareUserURLWithChatKey(contactID string) (string, error) {
 	publicKey, err := crypto.HexToPubkey(contactID)
 	if err != nil {
 		return "", err
@@ -422,8 +410,8 @@ func parseUserURLWithChatKey(urlData string) (*URLDataResponse, error) {
 	}, nil
 }
 
-func (m *Messenger) ShareUserURLWithENS(contactID string) (string, error) {
-	contact := m.GetContactByID(contactID)
+func (m *Service) ShareUserURLWithENS(contactID string) (string, error) {
+	contact := m.provider.GetContactByID(contactID)
 	if contact == nil {
 		return "", ErrContactNotFound
 	}
@@ -477,8 +465,8 @@ func (m *Messenger) prepareEncodedUserData(contact *contacts.Contact) (string, s
 	return encodedData, shortKey, nil
 }
 
-func (m *Messenger) ShareUserURLWithData(contactID string) (string, error) {
-	contact := m.GetContactByID(contactID)
+func (m *Service) ShareUserURLWithData(contactID string) (string, error) {
+	contact := m.provider.GetContactByID(contactID)
 	if contact == nil {
 		return "", ErrContactNotFound
 	}
@@ -489,139 +477,4 @@ func (m *Messenger) ShareUserURLWithData(contactID string) (string, error) {
 	}
 
 	return fmt.Sprintf("%s/u/%s#%s", baseShareURL, data, shortKey), nil
-}
-
-func parseUserURLWithData(data string, chatKey string) (*URLDataResponse, error) {
-	if data == "" {
-		return &URLDataResponse{
-			Contact: &ContactURLData{
-				DisplayName: "",
-				Description: "",
-				PublicKey:   chatKey,
-			},
-		}, nil
-	}
-	urlData, err := decodeDataURL(data)
-	if err != nil {
-		return nil, err
-	}
-
-	var urlDataProto protobuf.URLData
-	err = proto.Unmarshal(urlData, &urlDataProto)
-	if err != nil {
-		return nil, err
-	}
-
-	var userProto protobuf.User
-	err = proto.Unmarshal(urlDataProto.Content, &userProto)
-	if err != nil {
-		return nil, err
-	}
-
-	return &URLDataResponse{
-		Contact: &ContactURLData{
-			DisplayName: userProto.DisplayName,
-			Description: userProto.Description,
-			PublicKey:   chatKey,
-		},
-	}, nil
-}
-
-func IsStatusSharedURL(url string) bool {
-	return strings.HasPrefix(url, sharedURLUserPrefix) ||
-		strings.HasPrefix(url, sharedURLUserPrefixWithData) ||
-		strings.HasPrefix(url, sharedURLCommunityPrefix) ||
-		strings.HasPrefix(url, sharedURLCommunityPrefixWithData) ||
-		strings.HasPrefix(url, sharedURLChannelPrefixWithData)
-}
-
-func splitSharedURLData(data string) (string, string, error) {
-	const count = 2
-	contents := strings.SplitN(data, "#", count)
-	if len(contents) != count {
-		return "", "", fmt.Errorf("url should contain at least one `#` separator")
-	}
-	return contents[0], contents[1], nil
-}
-
-func ParseSharedURL(url string) (*URLDataResponse, error) {
-
-	if strings.HasPrefix(url, sharedURLUserPrefix) {
-		chatKey := strings.TrimPrefix(url, sharedURLUserPrefix)
-		if strings.HasPrefix(chatKey, "zQ3sh") {
-			return parseUserURLWithChatKey(chatKey)
-		}
-		return parseUserURLWithENS(chatKey)
-	}
-
-	if strings.HasPrefix(url, sharedURLUserPrefixWithData) {
-		trimmedURL := strings.TrimPrefix(url, sharedURLUserPrefixWithData)
-		encodedData, chatKey, err := splitSharedURLData(trimmedURL)
-		if err != nil {
-			return nil, err
-		}
-		return parseUserURLWithData(encodedData, chatKey)
-	}
-
-	if strings.HasPrefix(url, sharedURLCommunityPrefix) {
-		chatKey := strings.TrimPrefix(url, sharedURLCommunityPrefix)
-		return parseCommunityURLWithChatKey(chatKey)
-	}
-
-	if strings.HasPrefix(url, sharedURLCommunityPrefixWithData) {
-		trimmedURL := strings.TrimPrefix(url, sharedURLCommunityPrefixWithData)
-		encodedData, chatKey, err := splitSharedURLData(trimmedURL)
-		if err != nil {
-			return nil, err
-		}
-		return parseCommunityURLWithData(encodedData, chatKey)
-	}
-
-	if strings.HasPrefix(url, sharedURLChannelPrefixWithData) {
-		trimmedURL := strings.TrimPrefix(url, sharedURLChannelPrefixWithData)
-		encodedData, chatKey, err := splitSharedURLData(trimmedURL)
-		if err != nil {
-			return nil, err
-		}
-
-		if channelRegExp.MatchString(encodedData) {
-			return parseCommunityChannelURLWithChatKey(encodedData, chatKey)
-		}
-		return parseCommunityChannelURLWithData(encodedData, chatKey)
-	}
-
-	return nil, fmt.Errorf("not a status shared url")
-}
-
-func encodeDataURL(data []byte) (string, error) {
-	bb := bytes.NewBuffer([]byte{})
-	writer := brotli.NewWriter(bb)
-	_, err := writer.Write(data)
-	if err != nil {
-		return "", err
-	}
-
-	err = writer.Close()
-	if err != nil {
-		return "", err
-	}
-
-	return base64.URLEncoding.EncodeToString(bb.Bytes()), nil
-}
-
-func decodeDataURL(data string) ([]byte, error) {
-	decoded, err := base64.URLEncoding.DecodeString(data)
-	if err != nil {
-		return nil, err
-	}
-
-	output := make([]byte, 4096)
-	bb := bytes.NewBuffer(decoded)
-	reader := brotli.NewReader(bb)
-	n, err := reader.Read(output)
-	if err != nil {
-		return nil, err
-	}
-
-	return output[:n], nil
 }
